% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Get metadata about the weather and climate elements that are
defined for use in the Frost API}
\usage{
get_elements(client_id, ...)

get_elements(client_id,
             ids = NULL,
             names = NULL,
             descriptions = NULL,
             units = NULL,
             codeTables= NULL,
             statuses = NULL,
             calculationMethod = NULL,
             categories = NULL,
             timeOffsets = NULL,
             sensorLevels = NULL,
             oldElementCodes = NULL,
             oldUnits = NULL,
             cfStandardNames = NULL,
             cfCellMethods = NULL,
             cfUnits = NULL,
             cfVersions = NULL,
             fields = NULL,
             return_response = FALSE)
}
\arguments{
\item{client_id}{A string. The client ID to use to send requests to the
Frost API.}

\item{ids}{A character vector. The element IDs to get metadata for.}

\item{names}{A character vector. The element names to get metadata for.}

\item{descriptions}{A character vector. The descriptions to get metadata
for.}

\item{units}{A character vector. The units to get metadata for.}

\item{statuses}{A character vector. The statuses to get metadata for.}

\item{categories}{A character vector. The categories to get metadata for.}

\item{fields}{A character vector. A vector of the fields that should be
present in the response. If not set, then all fields will be retrieved.}

\item{return_response}{A logical. If set to \code{TRUE}, then the function
returns the response from the GET request. If set to \code{FALSE} (default),
then the function returns a dataframe of the content in the response to the
GET request.}

\item{codeTables}{A character vector. The code tables to get metadata for.}

\item{calculationMethod}{A string. The calculation method as a JSON filter.
Supports the following keys: baseNames, methods, innerMethods, periods,
innerPeriods, tresholds, methodDescriptions, innerMethodDescriptions,
methodUnits, and innerMethodUnits.}

\item{timeOffsets}{A character vector. The time offsets to get metadata for.}

\item{sensorLevels}{A string. The sensor levels to get metadata
for as a JSON filter. Supports the following keys: levelTypes, units,
defaultValues, and values.}

\item{oldElementCodes}{A character vector. The old MET Norway element codes
to get metadata for.}

\item{oldUnits}{A character vector. The old MET Norway units to get
metadata for.}

\item{cfStandardNames}{A character vector. The CF standard names to get
metadata for.}

\item{cfCellMethods}{A character vector. The CF cell methods to get
metadata for.}

\item{cfUnits}{A character vector. The CF units to get metadata for.}

\item{cfVersions}{A character vector. The CF versions to get metadata for.}
}
\value{
The function returns either a data frame of weather and climate
elements, or the response of the GET request for location resource in
the Frost API, depending on the value set for the \code{return_response}
argument.
}
\description{
\code{get_elements()} retrieves metadata about weather and
climate elements defined for use in the Frost API. The function requires a
client ID for the \code{client_id} argument at minimum. Use the optional
input arguments to filter the set of elements returned in the response.
The optional function arguments default to NULL, which translates to no
filter on the returned response from the Frost API.
}
\examples{
client.id <- "<YOUR CLIENT ID>"

# Get data for all elements
elements <- get_elements(client_id = client.id)

}
