% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get metadata for the location names defined in the Frost API}
\usage{
get_locations(client_id, ...)

get_locations(client_id,
              names = NULL,
              geometry = NULL,
              fields = NULL,
              return_response = FALSE)
}
\arguments{
\item{client_id}{A string. The client ID to use to send requests to the
Frost API.}

\item{names}{A character vector. The location names that you want
metadata for.}

\item{geometry}{A string. Get Frost API location names defined by a
specified geometry. Geometries are specified as either "nearest(POINT(...))"
or "POLYGON(...)" using well-known text representation for geometry (WKT).}

\item{fields}{A character vector. A vector of the fields that should be
present in the response. If not set, then all fields will be retrieved.}

\item{return_response}{A logical. If set to \code{TRUE}, then the function
returns the response from the GET request. If set to \code{FALSE} (default),
then the function returns a dataframe of the content in the response to the
GET request.}
}
\value{
The function returns either a data frame of locations, or the
response of the GET request for location resource in the Frost API,
depending on the value set for the \code{return_response} argument.
}
\description{
\code{get_locations()} retrieves metadata about location names
defined for use in the Frost API. The function requires a client ID for the
\code{client_id} argument at minimum. Use the optional input arguments to
filter the set of location names returned in the response. The optional
function arguments default to NULL, which translates to no filter on the returned
response from the Frost API.
}
\examples{
client.id <- "<YOUR CLIENT ID>"

# Get data for all sources
locations <- get_locations(client_id = client.id)

}
