% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_qualitycodes.R
\name{get_obs_qualitycodes}
\alias{get_obs_qualitycodes}
\title{Get metadata about existing quality flags for observations}
\usage{
get_obs(client_id, sources, reference_time, elements, ...)

get_obs(client_id,
        sources,
        reference_time,
        elements,
        maxage = NULL,
        limit = NULL,
        time_offsets = NULL,
        time_resolutions = NULL,
        time_series_ids = NULL,
        performance_categories = NULL,
        exposure_categories = NULL,
        qualities = NULL,
        levels = NULL,
        include_extra = NULL,
        fields = NULL,
        return_response = FALSE)
}
\arguments{
\item{client_id}{A string. The client ID to use to send requests to the Frost
API.}

\item{fields}{A character vector. The field to include in the response (i.e.
output). If this parameter is set, then only the specified field is
returned as a data frame. If not set, then all fields will be
returned in the response as a list. The options are "summarized" and
"details".}

\item{language}{A string. The language of the fields in the response. The
options are "en-US" (default), "nb-NO" (Norwegian, Bokmål), and "nn-NO"
(Norwegian, Nynorsk).}

\item{return_response}{A logical. If set to \code{TRUE}, then the function
returns the response from the GET request. If set to \code{FALSE}, then the
function returns a dataframe of the content in the response to the GET
request.}
}
\value{
The function returns either a data frame of historical weather
observations or the response of the GET request for the observations,
depending on the value set for the \code{return_response} argument.
}
\description{
\code{get_obs_qualitycodes()} provides a data frame with all
possible detail values given the quality service. The function requires
input for \code{client_id}. The other function arguments are optional, and default
to \code{NULL}, which means that the response from the API is not
filtered on these parameters.
}
\examples{
client.id <- "<YOUR CLIENT ID>"

# Get metadata for quality codes
qualitycodes_df <- get_obs_qualitycodes(client_id = client.id)

# Get the summarized metadata for quality codes
summarized_df <- get_obs_qualitycodes(client_id = client.id,
                                      fields = "summarized")

}
